requirements.txt
fastapi
uvicorn[standard]
aiosqlite
pyjwt
pytest
pytest-mock
pytest-asyncio
pytest-docker
coverage
asyncpg
aioredis
requests
httpx